+++ libraries/LTR390_DFR/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.1.1] - 2024-09-23
+- extend functionality and documentation
+- add return value in **setGain()** 0..4
+- add param check in **bool setMeasurement(resolution,time)**
+- refactored code for readability (e.g. magic numbers).
+- add **bool setUVsensitivity(float s)** and **float getUVsensitivity()()**
+- update keywords.txt (prepare)
+
+
 ## [0.1.0] - 2024-04-29
 - initial version